Successfully moved python files to py-server folder
